Provide version information on a per-target basis
First, this scopes the version information to individual targets, e.g. it uses the provided target's `CMAKE_SOURCE_DIR` and `CMAKE_BINARY_DIR` properties to source git information and decide upon an output directory for the generated header. Secondly, this change removes the need for using `target_link_libraries`. Instead, a simple `#include <version.h>` in a source file of the target is sufficient. Thirdly, this makes version information generation explicit as opposed to have it happen implicitly as side-effect of the `include()` statement.