Install prerequisites via GitHub actions
There are two reasons for this change: First of all, `chocolatey` on Windows imposes usage limits. These are reasonable, but at the same time can come as a surprise in the form of random build failures. Secondly, relying on the package manager may not provide a desired version of the prerequisites.